ATA PIO, dup3, fcntl, rw locks, misc changes
new features:
shell redirections, pipes

new kernel features:
umask
lseek with 64 bit offsets
O_CLOEXEC
O_CLOFORK
O_SYNC
krealloc
pipe2
24 bit VBE region swap

new libc features:
realloc
v/snprintf
readjustable malloc heap
swab
sync
pipe2

bug fixes:
ysh not blocking TSTP, TTOU, TTIN
ysh not correctly handling EOF
cat not using stdin
cat not checking write return value
ls not correctly reporting errors
xxd not correctly printing buffer

kernel bug fixes:
actually working RS/232 now
pipes incorrectly set writer count to 0 leading to EOF
kasserts without ;
ll macros changed to work on a single line
fixed pipe closing (wrong order)
unblocking pipe queues on file close
proper dup, dup2
wrong return on openat_inode leading to EROFS where ENOENT would be more
appropriate
pipes instantly returning EOF on 0 writers instead of flushing their
buffer
proper offset type in inode_t
buffer overruns in cpuid
idle_task init early so that the kernel can use sys_nanosleep
fixed spurious interrupt detection in the PIC
fixed potential hang in PS/2 interrupt handler (not waking driver)
fixed memory ordering in spinlocks
fixed syscall context save/change saving to &ctx instead of ctx
fixed potential interrupt disable on RS/232 interrupt handler
(incorrect EOI ordering)
fixed potential hang in RS/232 interrupt handler (not waking driver)

libc bug fixes:
memset actually properly memsetting
strto* incorrectly returning end
printf incorrectly doing long
*printf, *scanf incorrectly doing/not doing va_end

generic clean up and some refactoring everywhere

+++ libc/src/include/string.h
@@ -3,8 +3,8 @@
 #include <stddef.h>
 #include <stdint.h>
 
-void itoad(uint32_t num, char * out); // signed int
-void itoaud(uint32_t num, char * out); // unsigned int
+void itoad (uint64_t num, char * out); // signed int
+void itoaud(uint64_t num, char * out); // unsigned int
 
 void itoax(uint32_t num, char * out);
 void i64toax(uint64_t i, char * out);
